+++ CHANGELOG.md
@@ -24,6 +24,23 @@ All notable changes to Sourcegraph are documented in this file.
 - Fix a bug where using a repo search filter with the prefix `github.com` only searched for repos whose name starts with `github.com`, even though no `^` was specified in the search filter. (#4103)
 
 ## 3.4.0 (unreleased)
+## 3.4.1 (unreleased)
+
+### Added
+
+- You may now specify `DISABLE_CONFIG_UPDATES=true` on the management console to prevent updates to the critical configuration. This is useful when loading critical config via a file using `CRITICAL_CONFIG_FILE` on the frontend.
+
+### Changed
+
+- When `EXTSVC_CONFIG_FILE` or `SITE_CONFIG_FILE` are specified, updates to external services and the site config are now prevented.
+
+### Removed
+
+### Fixed
+
+- Fixed an issue where `EXTSVC_CONFIG_FILE` being specified would incorrectly cause a panic.
+
+## 3.4.0
 
 ### Added